+++ projects/clr/hipamd/include/hip/amd_detail/hip_api_trace.hpp
@@ -641,6 +641,7 @@ typedef hipError_t (*t_hipMipmappedArrayGetLevel)(hipArray_t* pLevelArray,
 typedef hipError_t (*t_hipModuleGetFunction)(hipFunction_t* function, hipModule_t module,
                                              const char* kname);
 typedef hipError_t (*t_hipModuleGetFunctionCount)(unsigned int* count, hipModule_t module);
+typedef hipError_t (*t_hipModuleGetLoadingMode)(hipModuleLoadingMode_t* mode);
 typedef hipError_t (*t_hipModuleGetGlobal)(hipDeviceptr_t* dptr, size_t* bytes, hipModule_t hmod,
                                            const char* name);
 typedef hipError_t (*t_hipModuleGetTexRef)(textureReference** texRef, hipModule_t hmod,
@@ -1671,6 +1672,7 @@ struct HipDispatchTable {
   t_hipMemPrefetchAsync_v2 hipMemPrefetchAsync_v2_fn;
   t_hipMemAdvise_v2 hipMemAdvise_v2_fn;
   t_hipStreamGetId hipStreamGetId_fn;
+  t_hipModuleGetLoadingMode hipModuleGetLoadingMode_fn;
 
   // HIP_RUNTIME_API_TABLE_STEP_VERSION = 15
   t_hipLibraryLoadData hipLibraryLoadData_fn;