+++ projects/clr/hipamd/src/hip_code_object.cpp
@@ -50,13 +50,17 @@ hipError_t DynCO::loadCodeObject(const char* fname, const void* image) {
   IHIP_RETURN_ONFAIL(fb_info_->BuildProgram(ihipGetDevice()));
 
   module_ = fb_info_->Module(device_id_);
+  std::string mod_loading_mode;
+  if (!flagIsDefault(HIP_MODULE_LOADING)) {
+      mod_loading_mode = HIP_MODULE_LOADING;
+  }
+  if (mod_loading_mode == "EAGER" || mod_loading_mode == "eager") {
+    // Define Global variables
+    IHIP_RETURN_ONFAIL(populateDynGlobalVars());
 
-  // Define Global variables
-  IHIP_RETURN_ONFAIL(populateDynGlobalVars());
-
-  // Define Global functions
-  IHIP_RETURN_ONFAIL(populateDynGlobalFuncs());
-
+    // Define Global functions
+    IHIP_RETURN_ONFAIL(populateDynGlobalFuncs());
+  }
   return hipSuccess;
 }
 
@@ -90,7 +94,8 @@ DynCO::~DynCO() {
     delete elem.second;
   }
   functions_.clear();
-
+  dyn_data_loaded_ = false;
+  dyn_func_loaded_ = false;
   delete fb_info_;
 }
 
@@ -113,7 +118,7 @@ hipError_t DynCO::getDynFunc(hipFunction_t* hfunc, std::string func_name) {
   if (hfunc == nullptr) {
     return hipErrorInvalidValue;
   }
-
+  IHIP_RETURN_ONFAIL(populateDynGlobalFuncs());
   auto it = functions_.find(func_name);
   if (it == functions_.end()) {
     LogPrintfError("Cannot find the function: %s ", func_name.c_str());
@@ -129,12 +134,14 @@ hipError_t DynCO::getFuncCount(unsigned int* count) {
   if (count == nullptr) {
     return hipErrorInvalidValue;
   }
+  IHIP_RETURN_ONFAIL(populateDynGlobalFuncs());
   *count = functions_.size();
   return hipSuccess;
 }
 
 bool DynCO::isValidDynFunc(const void* hfunc) {
   amd::ScopedLock lock(dclock_);
+  IHIP_RETURN_ONFAIL(populateDynGlobalFuncs());
   return std::any_of(functions_.begin(), functions_.end(),
                      [&](auto& it) { return it.second->isValidDynFunc(hfunc); });
 }
@@ -195,6 +202,9 @@ hipError_t DynCO::initDynManagedVars(const std::string& managedVar) {
 hipError_t DynCO::populateDynGlobalVars() {
   amd::ScopedLock lock(dclock_);
   hipError_t err = hipSuccess;
+  if (dyn_data_loaded_) {
+    return err;
+  }
   std::vector<std::string> var_names;
   std::string managedVarExt = ".managed";
   // For Dynamic Modules there is only one hipFatBinaryDevInfo_
@@ -218,12 +228,15 @@ hipError_t DynCO::populateDynGlobalVars() {
       err = initDynManagedVars(managedVar);
     }
   }
+  dyn_data_loaded_ = true;
   return err;
 }
 
 hipError_t DynCO::populateDynGlobalFuncs() {
   amd::ScopedLock lock(dclock_);
-
+  if(dyn_func_loaded_) {
+    return hipSuccess;
+  }
   std::vector<std::string> func_names;
   device::Program* dev_program = fb_info_->GetProgram(ihipGetDevice())
                                      ->getDeviceProgram(*hip::getCurrentDevice()->devices()[0]);
@@ -237,7 +250,7 @@ hipError_t DynCO::populateDynGlobalFuncs() {
   for (auto& elem : func_names) {
     functions_.insert(std::make_pair(elem, new Function(elem)));
   }
-
+  dyn_func_loaded_ = true;
   return hipSuccess;
 }