+++ projects/hip/include/hip/hip_runtime_api.h
@@ -274,6 +274,11 @@ typedef enum hipMemoryType {
 
 } hipMemoryType;
 
+typedef enum hipModuleLoadingMode_t {
+  HIP_MODULE_EAGER_LOADING = 1,
+  HIP_MODULE_LAZY_LOADING = 2
+} hipModuleLoadingMode_t;
+
 /**
  * Pointer attributes
  */
@@ -6410,6 +6415,16 @@ hipError_t hipLibraryGetKernel(hipKernel_t* pKernel, hipLibrary_t library, const
  * @return #hipSuccess, #hipErrorInvalidValue
 */
 hipError_t hipLibraryGetKernelCount(unsigned int *count, hipLibrary_t library);
+
+/**
+ * @brief Function gets the current module load mode
+ *
+ * @param [out] module  mode of current module load
+ *
+ * @returns #hipSuccess, #hipErrorInvalidValue, #hipErrorInvalidContext, #hipErrorNotInitialized,
+ * #hipErrorNotFound,
+ */
+hipError_t hipModuleGetLoadingMode(hipModuleLoadingMode_t* mode);
 
 /**
  * @brief Find out attributes for a given function.